Rename ows panels and components to addLayers
Split getcapabilities querying into seperate service.
Ows was too ambiguous name, so rename to addLayers.
Use component architecture instead of controller+directive.

refs #217
