+++ common/wms/get-capabilities.service.js
@@ -3,13 +3,14 @@ import 'utils';
 import { Tile } from 'ol/layer';
 import { TileWMS } from 'ol/source';
 import { Attribution } from 'ol/control.js';
+import { getPreferedFormat } from '../format-utils';
 
 export default ['$http', 'hs.map.service', 'hs.utils.service', '$rootScope', function ($http, OlMap, utils, $rootScope) {
     var me = this;
 
     /**
     * Get WMS service location without parameters from url string
-    * @memberof hs.ows.wms.service_capabilities
+    * @memberof hs.wms.getCapabilitiesService
     * @function getPathFromUrl
     * @param {String} str Url string to parse
     * @returns {String} WMS service Url
@@ -23,7 +24,7 @@ export default ['$http', 'hs.map.service', 'hs.utils.service', '$rootScope', fun
 
     /**
     * Create WMS parameter string from parameter object 
-    * @memberof hs.ows.wms.service_capabilities
+    * @memberof hs.wms.getCapabilitiesService
     * @function param2String
     * @param {Object} obj Object with stored WNS service parameters
     * @returns {String} Parameter string or empty string if no object given 
@@ -44,7 +45,7 @@ export default ['$http', 'hs.map.service', 'hs.utils.service', '$rootScope', fun
 
     /**
     * Parse added service url and sends GetCapabalities request to WMS service
-    * @memberof hs.ows.wms.service_capabilities
+    * @memberof hs.wms.getCapabilitiesService
     * @function requestGetCapabilities
     * @param {String} service_url Raw Url localization of service
     * @returns {Promise} Promise object - Response to GetCapabalities request
@@ -72,7 +73,7 @@ export default ['$http', 'hs.map.service', 'hs.utils.service', '$rootScope', fun
 
     /**
     * Load all layers of selected service to the map
-    * @memberof hs.ows.wms.service_capabilities
+    * @memberof hs.wms.getCapabilitiesService
     * @function service2layers
     * @param {String} capabilities_xml Xml response of GetCapabilities of selected service
     * @returns {Ol.collection} List of layers from service
@@ -126,7 +127,7 @@ export default ['$http', 'hs.map.service', 'hs.utils.service', '$rootScope', fun
 
     /**
     * Test if current map projection is in supported projection list
-    * @memberof hs.ows.wms.service_capabilities
+    * @memberof hs.wms.getCapabilitiesService
     * @function currentProjectionSupported
     * @param {Array} srss List of supported projections
     * @returns {Boolean} True if map projection is in list, otherwise false
