+++ common_paths.js
@@ -20,6 +20,7 @@ exports.paths = [
   __dirname + '/components/api',
   root + '/node_modules/angular-material-bottom-sheet-collapsible',
   root + '/node_modules/bootstrap/dist/js',
+  __dirname + '/components/common',
   __dirname + '/components/hscesium',
   __dirname + '/components/compositions',
   __dirname + '/components/core',
@@ -44,7 +45,7 @@ exports.paths = [
   __dirname + '/components/measure',
   __dirname + '/components/mobile_settings',
   __dirname + '/components/mobile_toolbar',
-  __dirname + '/components/ows',
+  __dirname + '/components/add-layers/',
   __dirname + '/components/permalink',
   __dirname + '/components/print',
   __dirname + '/components/query',
