+++ components/add-layers/partials/add-layers.directive.html
@@ -1,5 +1,5 @@
 <div class="hs-panel-margin panel" ng-class="{'panel-default mainpanel' : !Core.singleDatasources}">
-    <div class="card-header" ng-if="!Core.singleDatasources" translate>Add external data
+    <div class="card-header" ng-show="!Core.singleDatasources" translate>Add external data
         <button type="button" class="close" ng-click="Core.closePanel(this)">
             <span aria-hidden="true">×</span>
             <span class="sr-only" translate>Close</span>
