+++ components/add-layers/partials/add-vector-layer.directive.html
@@ -1,5 +1,5 @@
 <p><sub class="text-danger">*Note: if you link a file with coordinates in World Geodetic System (lon/lat), please set in the "ADVANCED OPTIONS" the SRS to EPSG:4326</sub></p>
-<form class="form-horizontal" role="form" ng-controller="hs.ows.nonwms.controller">
+<form class="form-horizontal" role="form" ng-controller="hs.addLayersVector.controller">
     <div class="form-group">
         <label class="capabilities_label col-sm-4 control-label" translate>Title</label>
         <div class="capabilities_input col-sm-8">
