+++ components/add-layers/partials/add-wms-layer.directive.html
@@ -1,4 +1,4 @@
-<form class="form-horizontal" role="form" ng-controller="hs.ows.wms.controller">
+<form class="form-horizontal" role="form">
 
     <div class="form-group">
         <div class="input-group">
