+++ components/add-layers/partials/add-wms-layer.md.directive.html
@@ -1,5 +1,4 @@
-<form ng-controller="hs.ows.wms.controller">
-
+<form>
     <div flex layout="row">
         <md-input-container>
             <label translate>Image format</label>
