+++ components/add-layers/partials/dialog_getcapabilities_error.html
