+++ components/add-layers/partials/dialog_proxyconfirm.html
