+++ components/layermanager/hs.layermanager.layerlistDirective.js
@@ -5,7 +5,7 @@
  * @description Layer manager module maintain management of layers loaded in HS Layers application. It use folder structure to enable building hiearchy of layers. All layers are wrapped inside HSLayer object, which contain auxilary informations and layer itself.
  */
 
-define(['angular', 'ol', 'hs.source.SparqlJson', 'angular-socialshare', 'map', 'ows_nonwms', 'config_parsers', 'hs.layermanager.WMSTservice'],
+define(['angular', 'ol', 'hs.source.SparqlJson', 'angular-socialshare', 'map', 'hs.layermanager.WMSTservice'],
 
     function (angular, ol, SparqlJson, social) {
         return {
