+++ components/layermanager/layermanager.js
@@ -9,7 +9,6 @@ import { TileWMS, WMTS, OSM } from 'ol/source';
 import {ImageWMS, ImageArcGISRest} from 'ol/source';
 import {  WMSCapabilities } from 'ol/format';
 import 'utils';
-import 'ows_wms';
 import 'angular-drag-and-drop-lists';
 import 'status_creator';
 import 'styles';
@@ -22,7 +21,7 @@ import 'legend';
  * @description Layer manager module maintain management of layers loaded in HS Layers application. It use folder structure to enable building hiearchy of layers. All layers are wrapped inside HSLayer object, which contain auxilary informations and layer itself.
  */
 
-angular.module('hs.layermanager', ['hs.map', 'hs.utils', 'hs.ows.wms', 'dndLists', 'hs.status_creator', 'hs.styles', 'hs.legend'])
+angular.module('hs.layermanager', ['hs.map', 'hs.utils', 'dndLists', 'hs.status_creator', 'hs.styles', 'hs.legend'])
 
     /**
      * @module hs.layermanager
