+++ components/layout/partials/bottom-sheet.html
@@ -14,7 +14,7 @@
   </md-toolbar>
 	<div hs.layermanager.directive          ng-controller="hs.layermanager.controller" ng-if="Core.exists('hs.layermanager.controller')" ng-show="Core.panelVisible('layermanager', this)"></div>
 	<hs.legend              								ng-if="Core.exists('hs.legend.controller')" ng-show="Core.panelVisible('legend', this)"></hs.legend>
-	<div hs.ows.directive                   ng-controller="hs.ows.controller" ng-if="Core.exists('hs.ows.controller') && !Core.singleDatasources" ng-show="Core.panelVisible('ows', this)"></div>
+	<div hs.addLayers                   ng-controller="hs.addLayers" ng-if="Core.exists('hs.addLayers') && !Core.singleDatasources" ng-show="Core.panelVisible('ows', this)"></div>
 	<div hs.datasource_selector.directive   ng-controller="hs.datasource_selector.controller" ng-if="Core.exists('hs.datasource_selector.controller')" ng-show="Core.panelVisible('datasource_selector', this)"></div>
 	<div hs.compositions.directive          ng-controller="hs.compositions.controller" ng-if="Core.exists('hs.compositions.controller')" ng-show="Core.panelVisible('composition_browser', this)"></div>
 	<div hs.lodexplorer.directive           ng-controller="hs.lodexplorer.controller" ng-if="Core.exists('hs.lodexplorer.controller')" ng-show="Core.panelVisible('lodexplorer', this)"></div>
